lab2/Fix Phong illumination specular component reflecting light color.

+++ lab2/Code/scene.h
@@ -35,7 +35,7 @@ class Scene
     private:
         Vector getNormalizedLightVectorFromPosition(Point position);
         double calculateDiffuseComponent(Vector normal, Point hit);
-        double calculateSpecularComponent(Vector normal, Point hit, int p);
+        double calculateSpecularComponent(Vector normal, Point hit, double p);
 };
 
 #endif
